# docs/assistants/specwriter/actions-openapi.yaml
openapi: 3.1.0
info:
  title: QuickLabel Actions API
  version: 0.1.0
  summary: SpecWriter/LogDoctor が参照・呼び出すための最小OpenAPI
servers:
  - url: https://example.local
paths:
  /api/dev/health:
    get:
      summary: Health check (guarded by ACTIONS_TOKEN)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  env: { type: [ 'string', 'null' ] }
                  openai: { type: boolean }
                  supabase: { type: boolean }
                  guarded: { type: boolean }
                  ts: { type: string, format: date-time }
  /api/diagnostics/runtime-logs:
    get:
      summary: Runtime logs (guarded by ACTIONS_TOKEN)
      parameters:
        - in: query
          name: since
          schema: { type: string, format: date-time }
        - in: query
          name: until
          schema: { type: string, format: date-time }
        - in: query
          name: jobId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [OK, ERROR] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        jobId: { type: string }
                        step: { type: string }
                        status: { type: string, enum: [OK, ERROR] }
                        cause: { type: [ 'string', 'null' ] }
                        created_at: { type: string }
                  nextCursor: { type: [ 'string', 'null' ] }
        '401': { description: Unauthorized }
        '422': { description: Validation error }
  /api/run-e2e:
    post:
      summary: Run E2E suite (guarded by ACTIONS_TOKEN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                suite: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string }
securitySchemes:
  bearerAuth:
    type: http
    scheme: bearer
    bearerFormat: JWT